chore: migrate supported configs to typescript
