+++ .prettierrc.ts
@@ -1,7 +1,11 @@
-module.exports = {
+import type { Config } from 'prettier';
+
+const config: Config = {
   arrowParens: 'always',
   singleQuote: true,
   jsxSingleQuote: true,
   tabWidth: 2,
   semi: true,
 };
+
+export default config;
