+++ next.config.js
@@ -1,18 +1,15 @@
 /** @type {import('next').NextConfig} */
 /* eslint-disable @typescript-eslint/no-var-requires */
 
-// 检测是否为 Cloudflare Pages 构建
 const isCloudflare =
   process.env.CF_PAGES === '1' || process.env.BUILD_TARGET === 'cloudflare';
 const isLintRun =
   process.argv.includes('lint') || process.env.NEXT_DISABLE_PWA === '1';
 
 const nextConfig = {
-  // Cloudflare Pages 不支持 standalone，使用默认输出
   output: isCloudflare ? undefined : 'standalone',
   eslint: {
     dirs: ['src'],
-    // 在生产构建时忽略 ESLint 错误
     ignoreDuringBuilds: true,
   },
 
@@ -23,7 +20,6 @@ const nextConfig = {
     instrumentationHook: process.env.NODE_ENV === 'production' && !isCloudflare,
   },
 
-  // Uncoment to add domain whitelist
   images: {
     unoptimized: true,
     remotePatterns: [
@@ -39,23 +35,20 @@ const nextConfig = {
   },
 
   webpack(config, { isServer }) {
-    // Grab the existing rule that handles SVG imports
     const fileLoaderRule = config.module.rules.find((rule) =>
       rule.test?.test?.('.svg'),
     );
 
     config.module.rules.push(
-      // Reapply the existing rule, but only for svg imports ending in ?url
       {
         ...fileLoaderRule,
         test: /\.svg$/i,
-        resourceQuery: /url/, // *.svg?url
+        resourceQuery: /url/,
       },
-      // Convert all other *.svg imports to React components
       {
         test: /\.svg$/i,
         issuer: { not: /\.(css|scss|sass)$/ },
-        resourceQuery: { not: /url/ }, // exclude if *.svg?url
+        resourceQuery: { not: /url/ },
         loader: '@svgr/webpack',
         options: {
           dimensions: false,
@@ -64,7 +57,6 @@ const nextConfig = {
       },
     );
 
-    // Modify the file loader rule to ignore *.svg, since we have it handled now.
     fileLoaderRule.exclude = /\.svg$/i;
 
     config.resolve.fallback = {
@@ -74,7 +66,6 @@ const nextConfig = {
       crypto: false,
     };
 
-    // Exclude better-sqlite3, D1, and Postgres modules from client-side bundle
     if (!isServer) {
       config.externals = config.externals || [];
       config.externals.push({
