+++ package.json
@@ -83,6 +83,7 @@
     "@commitlint/cli": "^16.3.0",
     "@commitlint/config-conventional": "^16.2.4",
     "@opennextjs/cloudflare": "^1.15.1",
+    "@rollup/pluginutils": "^5.3.0",
     "@svgr/webpack": "^8.1.0",
     "@tailwindcss/forms": "^0.5.10",
     "@tailwindcss/typography": "^0.5.19",
@@ -117,10 +118,11 @@
     "prettier": "^3.8.1",
     "prettier-plugin-tailwindcss": "^0.7.1",
     "tailwindcss": "^3.4.17",
+    "ts-node": "^10.9.2",
     "typescript": "^5.9.3",
     "vercel": "^50.4.10",
-    "workbox-build": "6.6.0",
-    "webpack-obfuscator": "^3.5.1"
+    "webpack-obfuscator": "^3.5.1",
+    "workbox-build": "6.5.4"
   },
   "lint-staged": {
     "**/*.{js,jsx,ts,tsx}": [
